Reduce number of purge requests to CloudFlare
When index.html is purged from a folder, then we also need to purge the folder URI
from CloudFlare because this is cached separately.

For example: /about/index.html is purged, then also purge /about/.

Originally each folder was purged in a separate request, now these are
batched with all the other files.

Also added tests to verify behavior.